backend/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2.6
django-cors-headers==3.1.1
djangorestframework==3.10.3
et-xmlfile==1.0.1
future==0.18.0
gunicorn==19.9.0
idna==2.8
jdcal==1.4.1
psycopg2==2.8.3
pytz==2019.3
ratelim==0.1.6
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==4.1.4
